+++ TwistedLogik.Ultraviolet/Graphics/Graphics2D/CharacterRegion.cs
@@ -105,7 +105,7 @@ namespace TwistedLogik.Ultraviolet.Graphics.Graphics2D
         /// </summary>
         public Int32 Count
         {
-            get { return 1 + (Int32)end - (Int32)start; }
+            get { return 1 + end - start; }
         }
 
         // Property values.
